+++ app/controllers/api/trainings_controller.rb
@@ -1,8 +1,8 @@
 class API::TrainingsController < API::ApiController
   include ApplicationHelper
 
-  before_action :authenticate_user!, except: [:index, :show]
-  before_action :set_training, only: [:update, :destroy]
+  before_action :authenticate_user!, except: %i[index show]
+  before_action :set_training, only: %i[update destroy]
 
   def index
     @requested_attributes = params[:requested_attributes]
@@ -12,7 +12,8 @@ class API::TrainingsController < API::ApiController
     end
 
     if attribute_requested?(@requested_attributes, 'availabilities')
-      @trainings = @trainings.includes(:availabilities => [:slots => [:reservation => [:user => [:profile, :trainings]]]]).order('availabilities.start_at DESC')
+      @trainings = @trainings.includes(availabilities: [slots: [reservation: [user: %i[profile trainings]]]])
+                             .order('availabilities.start_at DESC')
     end
   end
 
@@ -39,12 +40,10 @@ class API::TrainingsController < API::ApiController
       end
 
       head :no_content
+    elsif @training.update(training_params)
+      render :show, status: :ok, location: @training
     else
-      if @training.update(training_params)
-        render :show, status: :ok, location: @training
-      else
-        render json: @training.errors, status: :unprocessable_entity
-      end
+      render json: @training.errors, status: :unprocessable_entity
     end
   end
 
@@ -57,19 +56,24 @@ class API::TrainingsController < API::ApiController
   def availabilities
     authorize Training
     @training = Training.find(params[:id])
-    @availabilities = @training.availabilities.includes(slots: {reservations: {user: [:profile, :trainings] }}).order('start_at DESC')
+    @availabilities = @training.availabilities
+                               .includes(slots: { reservations: { user: %i[profile trainings] } })
+                               .order('start_at DESC')
   end
 
   private
-    def set_training
-      @training = Training.find(params[:id])
-    end
 
-    def valid_training_params
-      params.require(:training).permit(:id, users: [])
-    end
+  def set_training
+    @training = Training.find(params[:id])
+  end
 
-    def training_params
-      params.require(:training).permit(:id, :name, :description, :machine_ids, :plan_ids, :nb_total_places, :public_page, :disabled, training_image_attributes: [:attachment], machine_ids: [], plan_ids: [])
-    end
+  def valid_training_params
+    params.require(:training).permit(:id, users: [])
+  end
+
+  def training_params
+    params.require(:training)
+          .permit(:id, :name, :description, :machine_ids, :plan_ids, :nb_total_places, :public_page, :disabled,
+                  training_image_attributes: [:attachment], machine_ids: [], plan_ids: [])
+  end
 end