+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20181217110454) do
+ActiveRecord::Schema.define(version: 20190110150532) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -30,6 +30,15 @@ ActiveRecord::Schema.define(version: 20181217110454) do
   end
 
   add_index "abuses", ["signaled_type", "signaled_id"], name: "index_abuses_on_signaled_type_and_signaled_id", using: :btree
+
+  create_table "accounting_periods", force: :cascade do |t|
+    t.date     "start_at"
+    t.date     "end_at"
+    t.datetime "closed_at"
+    t.integer  "closed_by"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
 
   create_table "addresses", force: :cascade do |t|
     t.string   "address",        limit: 255
@@ -770,12 +779,12 @@ ActiveRecord::Schema.define(version: 20181217110454) do
   add_index "user_trainings", ["user_id"], name: "index_user_trainings_on_user_id", using: :btree
 
   create_table "users", force: :cascade do |t|
-    t.string   "email",                  limit: 255, default: "",    null: false
-    t.string   "encrypted_password",     limit: 255, default: "",    null: false
+    t.string   "email",                  limit: 255, default: "",   null: false
+    t.string   "encrypted_password",     limit: 255, default: "",   null: false
     t.string   "reset_password_token",   limit: 255
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",                      default: 0,     null: false
+    t.integer  "sign_in_count",                      default: 0,    null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip",     limit: 255
@@ -784,7 +793,7 @@ ActiveRecord::Schema.define(version: 20181217110454) do
     t.datetime "confirmed_at"
     t.datetime "confirmation_sent_at"
     t.string   "unconfirmed_email",      limit: 255
-    t.integer  "failed_attempts",                    default: 0,     null: false
+    t.integer  "failed_attempts",                    default: 0,    null: false
     t.string   "unlock_token",           limit: 255
     t.datetime "locked_at"
     t.datetime "created_at"
@@ -795,7 +804,6 @@ ActiveRecord::Schema.define(version: 20181217110454) do
     t.string   "username",               limit: 255
     t.string   "slug",                   limit: 255
     t.boolean  "is_active",                          default: true
-    t.boolean  "invoicing_disabled",                 default: false
     t.string   "provider"
     t.string   "uid"
     t.string   "auth_token"
@@ -856,6 +864,7 @@ ActiveRecord::Schema.define(version: 20181217110454) do
 
   add_index "wallets", ["user_id"], name: "index_wallets_on_user_id", using: :btree
 
+  add_foreign_key "accounting_periods", "users", column: "closed_by"
   add_foreign_key "availability_tags", "availabilities"
   add_foreign_key "availability_tags", "tags"
   add_foreign_key "event_price_categories", "events"