+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog Fab Manager
 
+- Fixed missing translations (EN & ES) and improved others (ES)
+
 ## v2.8.1 2019 January 02
 
 - Fix ES upgrade: when docker-compose file is using ${PWD}, the ES config volume is attached to the wrong container