+++ CHANGELOG.md
@@ -1,11 +1,10 @@
 # Changelog Fab Manager
 
-- Fixed missing translations (EN & ES) and improved others (ES)
-
 ## v2.8.1 2019 January 02
 
 - Fix ES upgrade: when docker-compose file is using ${PWD}, the ES config volume is attached to the wrong container
 - Fixed environment documentation references for external locales
+- Fixed missing translations (EN & ES) and improved others (ES)
 - Fix a bug: unable to fetch projects from OpenProjects (#126)
 - Fix a bug: unable to create or edit a plan