+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Fix a bug: error handling on password recovery
 - Fix a bug: error handling on machine attachment upload
+- Fix a bug: first day of week is ignored in statistics custom filter
+- Fix a bug: rails DSB locale is invalid
+- Refactored frontend invoices translations
 
 ## v2.8.1 2019 January 02